support configuring the filter bits per key
adds a new db opt named filter_bits_per_key which lets the user
specify the number of filter bits to use per key for each bloom
filter.

also changes the way that SsTableFormat is built to use a builder
rather than a new() method with a bunch of parameters. This will
let us add new format parameters without having to change all the
places where we create a format